+++ packages/hooks/stories/useResize.stories.js
@@ -1,6 +1,6 @@
-import React, { useEffect, useRef } from 'react';
+import React, { useEffect, useRef, useState } from 'react';
 
-import { Card } from '@elonwu/web';
+import { Title, Text, Card } from '@elonwu/web';
 import { useResize } from '../src';
 
 export default {
@@ -11,18 +11,51 @@ export const UseResizeStory = () => {
   const targetRef = useRef();
   const rect = useResize(targetRef);
 
-  console.log({ rect });
+  const [scale, setScale] = useState(1);
+
+  const onScale = () => {
+    setScale((prev) => {
+      const result = prev === 1 ? 1.2 : 1;
+      return result;
+    });
+  };
 
   return (
-    <div style={{ height: 200, overflow: 'auto', border: '1px solid #ededed' }}>
+    <div
+      style={{
+        width: '100vw',
+        height: '100vh',
+        display: 'grid',
+        placeContent: 'stretch',
+        placeItems: 'center',
+      }}
+    >
       <Card
         ref={targetRef}
         style={{
-          width: '50vw',
-          height: '50vh',
+          width: '50%',
+          height: '50%',
           background: '#f3f4f5',
+          display: 'grid',
+          placeContent: 'center',
+          position: 'relative',
+          transform: `scale(${scale})`,
+          transition: 'all .2s ease',
         }}
-      />
+        onClick={onScale}
+      >
+        <Text style={{ position: 'absolute', top: 8, left: 8 }}>
+          ({rect.left},{rect.top})
+        </Text>
+
+        <Text style={{ position: 'absolute', bottom: 8, right: 8 }}>
+          ({rect.right},{rect.bottom})
+        </Text>
+
+        <Title>
+          {rect.width}*{rect.height}
+        </Title>
+      </Card>
     </div>
   );
 };