+++ packages/web/menu/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.6](https://github.com/ElonWu/elonwu_ui/compare/@elonwu/web-menu@0.0.5...@elonwu/web-menu@0.0.6) (2021-10-18)
+
+**Note:** Version bump only for package @elonwu/web-menu
+
+
+
+
+
 ## [0.0.5](https://github.com/ElonWu/elonwu_ui/compare/@elonwu/web-menu@0.0.4...@elonwu/web-menu@0.0.5) (2021-10-14)
 
 **Note:** Version bump only for package @elonwu/web-menu