+++ packages/web/portal/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.8](https://github.com/ElonWu/elonwu_ui/compare/@elonwu/web-portal@0.0.7...@elonwu/web-portal@0.0.8) (2021-10-18)
+
+**Note:** Version bump only for package @elonwu/web-portal
+
+
+
+
+
 ## [0.0.7](https://github.com/ElonWu/elonwu_ui/compare/@elonwu/web-portal@0.0.6...@elonwu/web-portal@0.0.7) (2021-10-14)